refactor:
refactor: Update messages according to their conversation ignored

Pass the conversation ID to the backend and update messages associated
to that conversation instead of calculating which messages should have
their read status updated to true on the client side.

Protect route from users updating messages who are not participants in a
conversation.